Implemented Dynamic Information
Based on the result, we could conclude:
- The new dynamic feature selection process (SHAP-Based) has led to better generalization
- The model now adapts to market shifts, reducing reliance on less useful features

Model Performance Analysis:
1. Improve Loss & Generalization
- The training loss has steadily decreased from 0.117 -> 0.032 over 20 epochs
- This suggests us that the newly selected features(timestamp, closing_price, highest_price, Adj Close, Rolling_Min_5, Rolling_Max_5) and reduced noise had leads to a stable model performance
2. Evaluation Metrics
- MSE: 0.0327 (previously 0.0375) → Improvement
- RMSE: 0.1808 (previously 0.1937) → Lower Error
- R²: 0.9274 (previously 0.9168) → Higher variance explained
- MAPE: 1.2406 (previously 1.5595) → Better relative error reduction

Overall, the model is now making more accurate predictions with lower error and better variance capture. The reduction in MAPE suggests that it is improving in predicting %-based deviations which is essential for financial forecasting.

We also did compare the baseline model with XGBoost baseline:
1. XGBoost still gives an R² of 1.0000 with nearly perfect MSE (0.0000) -> overfitting
The LSTM model is improving steadily and generalizing better, while XGBoost might need a deeper evaluation on unseen data to verify true performance.